refactor: 🚚 Move generic symbols to a generic dir
