+++ modules/ppcp-sdk-v6/resources/js/methods/methodRegistry.test.js
