+++ modules/ppcp-sdk-v6/resources/js/methods/sessionPayment.js
@@ -1,8 +1,9 @@
 /**
- * Drives the create, confirm and approve sequence for a wallet payment.
+ * Drives the create, confirm and approve sequence for a merchant-presented
+ * payment.
  *
- * The wallet-specific parts (confirmOrder payload, mapped contact) arrive as
- * arguments, so Apple Pay can reuse the sequence. DOM-free: failures throw
+ * The method-specific parts (confirmOrder payload, mapped contact) arrive as
+ * arguments, so every method reuses the sequence. DOM-free: failures throw
  * and the caller reports them.
  *
  * @package
