+++ modules/ppcp-sdk-v6/resources/js/wallets/applePay.test.js
@@ -14,12 +14,12 @@ jest.mock( '../utils/errorHandler', () => ( {
 } ) );
 
 const mockResolveWalletTotal = jest.fn();
-jest.mock( './contextTotal', () => ( {
+jest.mock( '../methods/contextTotal', () => ( {
 	resolveContextTotal: ( ...args ) => mockResolveWalletTotal( ...args ),
 } ) );
 
 const mockPayWithWallet = jest.fn();
-jest.mock( './sessionPayment', () => ( {
+jest.mock( '../methods/sessionPayment', () => ( {
 	payWithSession: ( ...args ) => mockPayWithWallet( ...args ),
 } ) );
 
@@ -60,13 +60,13 @@ jest.mock( './applePayValidation', () => ( {
 } ) );
 
 const mockRevealWalletGateway = jest.fn();
-jest.mock( './gatewayPlacement', () => ( {
-	revealWalletGateway: ( ...args ) => mockRevealWalletGateway( ...args ),
+jest.mock( '../methods/gatewayPlacement', () => ( {
+	revealMethodGateway: ( ...args ) => mockRevealWalletGateway( ...args ),
 } ) );
 
 const mockWalletShippingRequired = jest.fn( () => false );
 const mockCreateShippingController = jest.fn();
-jest.mock( './methodShipping', () => ( {
+jest.mock( '../methods/methodShipping', () => ( {
 	methodShippingRequired: ( ...args ) =>
 		mockWalletShippingRequired( ...args ),
 	createShippingController: ( ...args ) =>
