+++ modules/ppcp-sdk-v6/resources/js/wallets/applePayShipping.js
@@ -7,7 +7,7 @@
  * @package
  */
 
-import { resolveOptionId } from './shippingQuote';
+import { resolveOptionId } from '../methods/shippingQuote';
 import { walletAddressToWc } from './walletContacts';
 
 /**
